backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai-whisper==20250625
torch>=2.0.0
numpy>=1.24.0
pytest==8.0.0
pytest-mock==3.12.0
pytest-cov==4.1.0